pyproject.toml
[tool.poetry]
name = "daav-sample-code-rag"
version = "0.1.0"
description = ""
authors = ["Daniel Ávila Vera <davera.017@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"
transformers = "^4.39.3"
torch = "^2.2.2"
llama-cpp-python = "^0.2.58"
langchain = "^0.1.14"
accelerate = "^0.28.0"
huggingface-hub = "0.20.3"
numexpr = "^2.10.0"
tiktoken = "^0.6.0"
datasets = "^2.18.0"
scipy = "^1.13.0"
sentence-transformers = "^2.6.1"
qdrant-client = "^1.8.2"
tqdm = "^4.66.2"
llama-index = "^0.10.28"
llama-index-readers-json = "^0.1.5"
llama-index-readers-file = "^0.1.16"
llama-index-embeddings-huggingface = "^0.2.0"
llama-index-vector-stores-qdrant = "^0.2.0"
llama-index-llms-llama-cpp = "^0.1.3"
llama-index-postprocessor-colbert-rerank = "^0.1.1"
llama-index-llms-langchain = "^0.1.3"
llama-index-retrievers-bm25 = "^0.1.3"
llama-index-readers-huggingface-fs = "^0.1.3"
llama-index-program-lmformatenforcer = "^0.1.2"
lm-format-enforcer = "^0.9.5"
pyvis = "^0.3.2"
ragas = "^0.1.7"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"
spacy = "^3.7.4"
hydra-core = "^1.3.2"
mlflow = "^2.12.1"
seaborn = "^0.13.2"
plotly = "^5.22.0"
nbformat = "^5.10.4"


[tool.poetry.group.dev.dependencies]
flake8-pyproject = "^1.2.3"
black = "^24.3.0"
isort = "^5.13.2"
flake8 = "^7.0.0"
matplotlib = "^3.8.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


# Linters
[tool.flake8]
ignore = ["E226","E302","E41", "E402", "E128"]
max-line-length = 120
exclude = ["tests/*"]
# Use extend-ignore to add to already ignored checks which are anti-patterns like W503.
extend-ignore = ["W503"]

[tool.black]
line-length = 120
target-version = ['py37', 'py38', 'py39', 'py310']

[tool.isort]
atomic = true
profile = "black"
line_length = 120
skip_gitignore = true
known_first_party = ["black", "blib2to3", "blackd", "_black_version"]
